requirements.txt
affine==2.3.1
asgiref==3.5.2
attrs==22.1.0
branca==0.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
Django==4.1.2
djangorestframework==3.14.0
Fiona==1.8.22
folium==0.13.0
fonttools==4.38.0
geopandas==0.12.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
munch==2.5.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
plotly==5.11.0
pyparsing==3.0.9
pyproj==3.4.0
pystac==1.6.1
pystac-client==0.5.1
python-dateutil==2.8.2
pytz==2022.5
rasterio==1.3.3
requests==2.28.1
rioxarray==0.12.3
Shapely==1.8.5.post1
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.3
tenacity==8.1.0
tzdata==2022.5
urllib3==1.26.12
xarray==2022.10.0
